Queue_implementation
This is a simple queue implementation using array. Queue works in LIFO/FIFO method. That means, whenever we have to dequeue element we have to remove it from the front index of the array. And whenever we are going to enqueue an element we have to insert it the end index of the array.